Fixes for 32-bit build of fuzzers
This change resolves some type-related issues that were leading to
loss-of-precision warnings when compiling for i386 in OSS-Fuzz.

Reviewed-on: https://dawn-review.googlesource.com/c/tint/+/66000
Auto-Submit: Alastair Donaldson <afdx@google.com>
Kokoro: Kokoro <noreply+kokoro@google.com>
Commit-Queue: Ryan Harrison <rharrison@chromium.org>

+++ fuzzers/tint_spirv_tools_fuzzer/fuzzer.cc
@@ -70,7 +70,8 @@ std::unique_ptr<Mutator> CreateMutator(const std::vector<uint32_t>& binary,
 
   assert(!types.empty() && "At least one mutator type must be specified");
   RandomGenerator generator(seed);
-  auto mutator_type = types[generator.GetUInt64(types.size())];
+  auto mutator_type =
+      types[generator.GetUInt32(static_cast<uint32_t>(types.size()))];
 
   const auto& mutator_params = context->params.mutator_params;
   switch (mutator_type) {
